+++ Back/ns_local_portal/Views/user.py
@@ -23,22 +23,22 @@ from sqlalchemy import create_engine
 import json
 
 
-@view_config(route_name='core/user',renderer='json',permission=NO_PERMISSION_REQUIRED )
-def getUsers(request):
-    """Return the list of all the users with their roles.
-    """
+# @view_config(route_name='core/user',renderer='json',permission=NO_PERMISSION_REQUIRED )
+# def getUsers(request):
+#     """Return the list of all the users with their roles.
+#     """
     
-	q = (session.query(T_User, T_Role)
-        .join(T_Role)
-        .filter(T_User.Fk_Trole == T_Role.ID_Role)
-        ).all()
+# 	q = (session.query(T_User, T_Role)
+#         .join(T_Role)
+#         .filter(T_User.Fk_Trole == T_Role.ID_Role)
+#         ).all()
 
-	q = q.columns(T_User.id, T_User.firstname, T_User.familyname, T_User.login, T_Role.nomRole)
+# 	q = q.columns(T_User.id, T_User.firstname, T_User.familyname, T_User.login, T_Role.nomRole)
 
 
-	results = request.dbsession.execute(q).fetchall()
-	data = [dict(row) for row in results]
-	return data
+# 	results = request.dbsession.execute(q).fetchall()
+# 	data = [dict(row) for row in results]
+# 	return data
     
 @view_config(
     route_name='core/currentUser',
